+++ packages/react-zeugma/src/entities/zeugma/model/useZeugmaContext.ts
@@ -25,5 +25,8 @@ export const useZeugmaContext = (): ZeugmaController => {
     findTabById: state.findTabById,
     getTabMetadata: state.getTabMetadata,
     getActiveTabMetadata: state.getActiveTabMetadata,
+    poppedOutTabIds: state.poppedOutTabIds,
+    popoutTab: actions.popoutTab,
+    dockTab: actions.dockTab,
   }
 }